% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_legend_outside.R
\name{add_legend_outside}
\alias{add_legend_outside}
\title{add_legend_outside}
\usage{
add_legend_outside(...)
}
\description{
add a legend outside the plot. taken from:
http://stackoverflow.com/questions/3932038/plot-a-legend-outside-of-the-plotting-area-in-base-graphics
}
\examples{
par(mar = c(5, 4, 1.4, 0.2))
plot(rnorm(50), rnorm(50), col=c("steelblue", "indianred"), pch=20)
add_legend_outside("topright", legend=c("Foo", "Bar"), pch=20, 
                  col=c("steelblue", "indianred"),
                  horiz=TRUE, bty='n', cex=0.8)
}
