% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by.df.R
\name{by.df}
\alias{by.df}
\title{by.df}
\usage{
\method{by}{df}(...)
}
\description{
Same as df, but Value is a data.frame, not a list.
}
\details{
Come df, ma value e' un data.frame e non una list
}
\examples{

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (...) 
{
    rr = by(...)
    attr(rr, "class") = "list"
    rr = data.frame(rr)
    rr = t(rr)
    if (!is.null(list(...)$INDICES)) {
        rownames(rr) = unique(list(...)$INDICES)
    }
    else rownames(rr) = unique(list(...)[[2]])
    rr
  }

}
