% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_xtable_totale.R
\name{freq_xtable_totale}
\alias{freq_xtable_totale}
\title{freq_xtable_totale}
\usage{
freq_xtable_totale(tab, digits = NULL, ...)
}
\arguments{
\item{tab}{a freq table}

\item{digits}{as in xtable}
}
\description{
as xtable, but put a line between befor the total. Also substitute `Sum` with `Totale`
}
