% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtable_with_dimnames.R
\name{print_xtable_with_dimnames}
\alias{print_xtable_with_dimnames}
\title{print_xtable_with_dimnames}
\usage{
print_xtable_with_dimnames(x, names_rows_cols = c("Rows", "Cols"),
  hline.after = getOption("xtable.hline.after", c(-1, 0, nrow(x))), ...)
}
\arguments{
\item{x}{as in xtable}
}
\description{
print_xtable_with_dimnames
}
