% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_effects_individual.R
\name{plot_effects_individual}
\alias{plot_effects_individual}
\title{plot_effects_individual}
\usage{
plot_effects_individual(data, pred_name, resp_name, predict_funct,
  col.by = NULL, npoints = 10, center_effs = TRUE)
}
\arguments{
\item{predict_funct}{a function that makes predictions}

\item{npoints}{10}
}
\examples{
n=50
x=matrix(rnorm(n*3),n,3)
x[,2]=sign(x[,2])
y=rnorm(n,X[,1]+X[,1]*X[,2])
x=data.frame(x)
x$y=y
mod=lm(y~X1*X2+X3,data=x)
summary(mod)

predict_funct=function(newdata) predict(mod,newdata=newdata)
plot_effects_individual(x,"X1","y",predict_funct=predict_funct)
plot_effects_individual(x,"X1","y",predict_funct=predict_funct,col.by = x$X2)
plot_effects_individual(x,"X3","y",predict_funct=predict_funct,center_effs = TRUE)
plot_effects_individual(x,"X3","y",predict_funct=predict_funct,center_effs = FALSE)

}
\author{
livio finos
}
